Move WeightHosts out into its own class
