+++ atlasdb-cassandra-integration-tests/src/test/java/com/palantir/atlasdb/keyvalue/cassandra/CassandraClientPoolIntegrationTest.java
@@ -16,28 +16,23 @@
 package com.palantir.atlasdb.keyvalue.cassandra;
 
 import static org.hamcrest.Matchers.is;
-import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
 import java.net.InetSocketAddress;
 import java.nio.charset.StandardCharsets;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
-import java.util.NavigableMap;
 
 import org.apache.cassandra.thrift.KsDef;
 import org.apache.cassandra.thrift.TokenRange;
 import org.apache.thrift.TException;
-import org.hamcrest.Matchers;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Test;
-import org.mockito.Mockito;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
@@ -176,121 +171,4 @@ public class CassandraClientPoolIntegrationTest {
 
     private FunctionCheckedException<CassandraClient, List<TokenRange>, Exception> describeRing =
             client -> client.rawClient().describe_ring("atlasdb");
-
-    @Test
-    public void testWeightedHostsWithUniformActivity() {
-        Map<InetSocketAddress, CassandraClientPoolingContainer> pools = ImmutableMap.of(
-                new InetSocketAddress(0), createMockClientPoolingContainerWithUtilization(10),
-                new InetSocketAddress(1), createMockClientPoolingContainerWithUtilization(10),
-                new InetSocketAddress(2), createMockClientPoolingContainerWithUtilization(10));
-
-        NavigableMap<Integer, InetSocketAddress> result = CassandraClientPoolImpl.WeightedHosts.create(pools).hosts;
-
-        int expectedWeight = result.firstEntry().getKey();
-        int prevKey = 0;
-        for (Map.Entry<Integer, InetSocketAddress> entry : result.entrySet()) {
-            int currWeight = entry.getKey() - prevKey;
-            assertEquals(expectedWeight, currWeight);
-            prevKey = entry.getKey();
-        }
-    }
-
-    @Test
-    public void testWeightedHostsWithLowActivityPool() {
-        InetSocketAddress lowActivityHost = new InetSocketAddress(2);
-        Map<InetSocketAddress, CassandraClientPoolingContainer> pools = ImmutableMap.of(
-                new InetSocketAddress(0), createMockClientPoolingContainerWithUtilization(10),
-                new InetSocketAddress(1), createMockClientPoolingContainerWithUtilization(10),
-                lowActivityHost, createMockClientPoolingContainerWithUtilization(0));
-
-        NavigableMap<Integer, InetSocketAddress> result = CassandraClientPoolImpl.WeightedHosts.create(pools).hosts;
-
-        int largestWeight = result.firstEntry().getKey();
-        InetSocketAddress hostWithLargestWeight = result.firstEntry().getValue();
-        int prevKey = 0;
-        for (Map.Entry<Integer, InetSocketAddress> entry : result.entrySet()) {
-            int currWeight = entry.getKey() - prevKey;
-            prevKey = entry.getKey();
-            if (currWeight > largestWeight) {
-                largestWeight = currWeight;
-                hostWithLargestWeight = entry.getValue();
-            }
-        }
-        assertEquals(lowActivityHost, hostWithLargestWeight);
-    }
-
-    @Test
-    public void testWeightedHostsWithMaxActivityPool() {
-        InetSocketAddress highActivityHost = new InetSocketAddress(2);
-        Map<InetSocketAddress, CassandraClientPoolingContainer> pools = ImmutableMap.of(
-                new InetSocketAddress(0), createMockClientPoolingContainerWithUtilization(5),
-                new InetSocketAddress(1), createMockClientPoolingContainerWithUtilization(5),
-                highActivityHost, createMockClientPoolingContainerWithUtilization(20));
-
-        NavigableMap<Integer, InetSocketAddress> result = CassandraClientPoolImpl.WeightedHosts.create(pools).hosts;
-
-        int smallestWeight = result.firstEntry().getKey();
-        InetSocketAddress hostWithSmallestWeight = result.firstEntry().getValue();
-        int prevKey = 0;
-        for (Map.Entry<Integer, InetSocketAddress> entry : result.entrySet()) {
-            int currWeight = entry.getKey() - prevKey;
-            prevKey = entry.getKey();
-            if (currWeight < smallestWeight) {
-                smallestWeight = currWeight;
-                hostWithSmallestWeight = entry.getValue();
-            }
-        }
-        assertEquals(highActivityHost, hostWithSmallestWeight);
-    }
-
-    @Test
-    public void testWeightedHostsWithNonZeroWeights() {
-        Map<InetSocketAddress, CassandraClientPoolingContainer> pools = ImmutableMap.of(
-                new InetSocketAddress(0), createMockClientPoolingContainerWithUtilization(5),
-                new InetSocketAddress(1), createMockClientPoolingContainerWithUtilization(10),
-                new InetSocketAddress(2), createMockClientPoolingContainerWithUtilization(15));
-
-        NavigableMap<Integer, InetSocketAddress> result = CassandraClientPoolImpl.WeightedHosts.create(pools).hosts;
-
-        int prevKey = 0;
-        for (Map.Entry<Integer, InetSocketAddress> entry : result.entrySet()) {
-            int currWeight = entry.getKey() - prevKey;
-            assertThat(currWeight, Matchers.greaterThan(0));
-            prevKey = entry.getKey();
-        }
-    }
-
-    // Covers a bug where we used ceilingEntry instead of higherEntry
-    @Test
-    public void testSelectingHostFromWeightedHostsMatchesWeight() {
-        Map<InetSocketAddress, CassandraClientPoolingContainer> pools = ImmutableMap.of(
-                new InetSocketAddress(0), createMockClientPoolingContainerWithUtilization(5),
-                new InetSocketAddress(1), createMockClientPoolingContainerWithUtilization(10),
-                new InetSocketAddress(2), createMockClientPoolingContainerWithUtilization(15));
-        CassandraClientPoolImpl.WeightedHosts weightedHosts = CassandraClientPoolImpl.WeightedHosts.create(pools);
-        Map<InetSocketAddress, Integer> hostsToWeight = new HashMap<>();
-        int prevKey = 0;
-        for (Map.Entry<Integer, InetSocketAddress> entry : weightedHosts.hosts.entrySet()) {
-            hostsToWeight.put(entry.getValue(), entry.getKey() - prevKey);
-            prevKey = entry.getKey();
-        }
-
-        // Exhaustively test all indexes
-        Map<InetSocketAddress, Integer> numTimesSelected = new HashMap<>();
-        for (int index = 0; index < weightedHosts.hosts.lastKey(); index++) {
-            InetSocketAddress host = weightedHosts.getRandomHostInternal(index);
-            if (!numTimesSelected.containsKey(host)) {
-                numTimesSelected.put(host, 0);
-            }
-            numTimesSelected.put(host, numTimesSelected.get(host) + 1);
-        }
-
-        assertEquals(hostsToWeight, numTimesSelected);
-    }
-
-    private static CassandraClientPoolingContainer createMockClientPoolingContainerWithUtilization(int utilization) {
-        CassandraClientPoolingContainer mock = Mockito.mock(CassandraClientPoolingContainer.class);
-        Mockito.when(mock.getOpenRequests()).thenReturn(utilization);
-        return mock;
-    }
 }
